Cleanup implementation, move cache to per-port.
